FIxed gcc warning.
git-svn-id: file:///home/survex-svn/survex/trunk@929 4b37db11-9a0c-4f06-9ece-9ab7cdaee568

+++ src/message.c
@@ -592,7 +592,7 @@ msg(int en)
 {
    /* NB can't use ASSERT here! */
    static char badbuf[256];
-   if (en >= 1000 && en < 1000 + sizeof(dontextract)/sizeof(char*))
+   if (en >= 1000 && en < 1000 + (int)(sizeof(dontextract)/sizeof(char*)))
       return dontextract[en - 1000];
    if (!msg_array) {
       if (en != 1)  {
